Operation-ize FindAndModify
